Added filtering functionality

+++ app/Http/Controllers/ListingController.php
@@ -2,20 +2,51 @@
 
 namespace App\Http\Controllers;
 
+use Inertia\Inertia;
 use App\Models\Listing;
 use Illuminate\Http\Request;
-use Inertia\Inertia;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class ListingController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
+        Gate::authorize('viewAny', Listing::class);
+
+        $filters = $request->only([
+            'priceFrom', 'priceTo', 'beds', 'baths', 'areaFrom', 'areaTo'
+        ]);
+
+        $query = Listing::orderByDesc('created_at')
+            ->when(
+                $filters['priceFrom'] ?? false,
+                fn ($query, $value) => $query->where('price', '>=', $value),
+             )->when(
+                $filters['priceTo'] ?? false,
+                fn ($query, $value) => $query->where('price', '<=', $value),
+             )->when(
+                $filters['beds'] ?? false,
+                fn ($query, $value) => $query->where('beds', (int)$value < 6 ? '=' : '>=', $value),
+             )->when(
+                $filters['baths'] ?? false,
+                fn ($query, $value) => $query->where('baths', (int)$value < 6 ? '=' : '>=', $value)
+             )->when(
+                $filters['areaFrom'] ?? false,
+                fn ($query, $value) => $query->where('area', '>=', $value)
+             )->when(
+                $filters['areaTo'] ?? false,
+                fn ($query, $value) => $query->where('area', '<=', $value),
+             );
+
         return inertia('Listing/Index',
         [
-            'listings' => Listing::all()
+            'filters' => $filters,
+            'listings' => $query->paginate(10)
+                ->withQueryString()
         ]);
     }
 
@@ -24,6 +55,7 @@ class ListingController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create', Listing::class);
         return inertia('Listing/Create');
     }
 
@@ -42,7 +74,6 @@ class ListingController extends Controller
                 'street' => 'required',
                 'street_nr' => 'required|min:1',
                 'price' => 'required|integer|min:10000'
-    
             ])
         );
 
@@ -54,6 +85,9 @@ class ListingController extends Controller
      */
     public function show(Listing $listing)
     {
+
+        Gate::authorize('view', $listing);
+
         return inertia('Listing/Show',
         [
             'listing' => $listing
@@ -65,6 +99,8 @@ class ListingController extends Controller
      */
     public function edit(Listing $listing)
     {
+        Gate::authorize('update', $listing);
+
         return inertia('Listing/Edit', [
             'listing' => $listing
         ]);
@@ -97,6 +133,8 @@ class ListingController extends Controller
      */
     public function destroy(Listing $listing)
     {
+        Gate::authorize('update', $listing);
+
         $listing->delete();
 
         return redirect()->back()->with('success', 'Listing was deleted!');
